Setup AWS S3 and MongoDB
Webp format images are stored in an AWS S3 bucket, the url pathway to them are stored in MongoDB, Mongo is automatically populated with the pathways to these images. Currently the test images in S3 are the strawhat pirates. Tested multiple times to make sure the images were retrievable. Only backend is setup , not front end.